GEOS-31 Revert previous commit.

+++ geo2seis/seismic_regridding.cpp
@@ -303,7 +303,6 @@ void SeismicRegridding::WriteParametersSegyInParallel(SeismicParameters         
   }
 }
 
-
 void SeismicRegridding::GenerateParameterGridForOutputQueue(GenResamplParam *params,
                                                             ResamplOutput   *resampl_output)
 {
@@ -895,9 +894,8 @@ void SeismicRegridding::FindVp(SeismicParameters &seismic_parameters, size_t n_t
       }
     }
   }
-
     ////-------------find edges---------------------
-  for (int k = topk; k <= botk + 1; k++) {
+  for (size_t k = topk; k <= botk + 1; k++) {
     for (size_t i = 0; i < egrid.GetNI() - 1; i++) {
       //bot edge
       size_t j = 0;
